runtime: play around with a `localize_structure` function
This function basically walks an object, looking for values which
look like handles. If one is found and is valid, it pushes it to a
queue and localizes the entire datastructure to a certain depth.

This seems to work for smoke test structures, but I plan on testing
with other more complicated structures in redis later.